视频id解析错误修复

+++ src/main/java/github/zimoyin/core/fanju/pojo/info/seriesI/Result.java
@@ -80,7 +80,7 @@ public class Result {
     /**
      * 番剧ssid
      */
-    private int season_id;
+    private long season_id;
     /**
      * 剧集标题
      */
